Fix early casting bug in the RMSNorm fused with add Triton Kernel

+++ aiter/ops/triton/rmsnorm.py
@@ -173,15 +173,16 @@ def _fused_add_rmsnorm_kernel(
                 cols = blk_idx * BLOCK_SIZE + col_offsets
                 input_ptrs = row_input_ptr + cols
                 input_ptrs = tl.multiple_of(input_ptrs, (16,))
-                x = tl.load(input_ptrs).to(tl.float32)
+                x = tl.load(input_ptrs)
                 res_in_ptrs = row_res_in_ptr + cols
                 res_in_ptrs = tl.multiple_of(res_in_ptrs, (16,))
-                res_in = tl.load(res_in_ptrs).to(tl.float32)
+                res_in = tl.load(res_in_ptrs)
                 x += res_in
                 # Stores residual_out
                 res_out_ptrs = row_res_out_ptr + cols
                 tl.store(res_out_ptrs, x.to(res_out_ptr.type.element_ty))
 
+                x = x.to(tl.float32)
                 sum_squares += tl.sum(x * x, axis=0)
 
             # Handle remainder
@@ -189,19 +190,16 @@ def _fused_add_rmsnorm_kernel(
             mask = cols < n_cols
             input_ptrs = row_input_ptr + cols
             input_ptrs = tl.multiple_of(input_ptrs, (16,))
-            x = tl.load(input_ptrs, mask=mask, other=0.0, cache_modifier=".cg").to(
-                tl.float32
-            )
+            x = tl.load(input_ptrs, mask=mask, other=0.0, cache_modifier=".cg")
             res_in_ptrs = row_res_in_ptr + cols
             res_in_ptrs = tl.multiple_of(res_in_ptrs, (16,))
-            res_in = tl.load(
-                res_in_ptrs, mask=mask, other=0.0, cache_modifier=".cg"
-            ).to(tl.float32)
+            res_in = tl.load(res_in_ptrs, mask=mask, other=0.0, cache_modifier=".cg")
             x += res_in
             # Stores residual_out
             res_out_ptrs = row_res_out_ptr + cols
             tl.store(res_out_ptrs, x.to(res_out_ptr.type.element_ty), mask=mask)
 
+            x = x.to(tl.float32)
             sum_squares += tl.sum(x * x, axis=0)
 
             # Compute normalization factor
@@ -238,19 +236,16 @@ def _fused_add_rmsnorm_kernel(
         for row_idx in tl.range(row_start, n_rows, NUM_PRGMS, num_stages=2):
             input_ptrs = input_ptr + row_idx * input_row_stride + col_offsets
             input_ptrs = tl.multiple_of(input_ptrs, (16,))
-            row = tl.load(input_ptrs, mask=mask, other=0.0, cache_modifier=".cg").to(
-                tl.float32
-            )
+            row = tl.load(input_ptrs, mask=mask, other=0.0, cache_modifier=".cg")
             res_in_ptrs = res_in_ptr + row_idx * input_row_stride + col_offsets
             res_in_ptrs = tl.multiple_of(res_in_ptrs, (16,))
-            res_in = tl.load(
-                res_in_ptrs, mask=mask, other=0.0, cache_modifier=".cg"
-            ).to(tl.float32)
+            res_in = tl.load(res_in_ptrs, mask=mask, other=0.0, cache_modifier=".cg")
             row += res_in
             # Stores residual_out
             res_out_ptrs = res_out_ptr + row_idx * input_row_stride + col_offsets
             res_out_ptrs = tl.multiple_of(res_out_ptrs, (16,))
             tl.store(res_out_ptrs, row.to(res_out_ptr.type.element_ty), mask=mask)
+            row = row.to(tl.float32)
 
             g = tl.load(g_ptr + col_offsets, mask=mask, other=0.0).to(tl.float32)
             row_norm = row * row
